==> temas/Guides/examples_rproject/01-tipos-datos.Rmd <==
---
title: "Tipos de datos en R"
author: "Edimer David Jaramillo"
date: "25/3/2021"
output: 
  html_document:
    toc: yes
    toc_float: yes
    number_sections: yes
    theme: cerulean
    highlight: pygments
    df_print: paged
    code_folding: show
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE)
```

# Tipos de datos en R

## Numéricos

-   En R los datos de tipo numérico se identifican como **numeric o double:**

```{r}
numero <- 34.80
numero
```

-   Podemos consultar el tipo de dato a través de la función **class():**

```{r}
class(numero)
```

- Vamos a crear un vector de números decimales (**numeric o double**):

```{r}
vector_numeros <- c(10.2, 20.5, 30.8)
vector_numeros
```

- Podemos preguntar la clase o tipo de dato del vector anterior:

```{r}
class(vector_numeros)
```

## Enteros

- Los enteros en R se tipifican como **integer**:

```{r}
enteros1 <- 5
class(enteros1)
```

- Todos los números en R (enteros o decimales) se tipifican como **numeric**, sin embargo, podemos forzar a R para que entienda un entero como tal:

```{r}
enteros2 <- 5L
class(enteros2)
```

- Puedo crear un vector de enteros:

```{r}
vector_enteros <- c(1L, 2L, 3L)
class(vector_enteros)
```

## Caracteres (texto)

- Los caracteres representan variables cualitativas, en R todo texto debe estar escrito entre comillas ("" o ''):

```{r}
texto <- "universidad"
class(texto)
```

- Podemos crear un vector de caracteres:

```{r}
vector_texto <- c("a", "b", "c")
class(vector_texto)
```

## Factores 

- Los factores representan variables cualitativas ordinales y se pueden crear dentro de R con la función **factor()**:

```{r}
vector_mes_malo <- factor(c("enero", "febrero", "marzo", "abril", "enero"))
class(vector_mes_malo)
```

- Podemos solicitar a R los niveles del factor a través de la función **levels()**:

```{r}
levels(vector_mes_malo)
```

- Podemos controlar el orden de los niveles en la función **factor()** editando el argumento **levels**:

```{r}
vector_mes_bueno <- factor(c("enero", "febrero", "marzo", "abril", "enero"),
                           levels = c("enero", "febrero", "marzo", "abril"))
levels(vector_mes_bueno)
```

## Lógicos

- En R los datos lógicos representan TRUE o FALSE:

```{r}
logico <- TRUE
class(logico)
```

- Podemos crear un vector de lógicos:

```{r}
vector_logicos <- c(TRUE, FALSE, TRUE, FALSE)
class(vector_logicos)
```

# Valores ausentes

- En R los valores ausentes se tipifican como **NA**:

```{r}
valor_ausente <- NA
class(valor_ausente)
```

- Un vecto puede incluir valores ausentes:

```{r}
vector_ausente <- c(34.6, NA, 56.9)
class(vector_ausente)
```

# Operadores

## Aritméticos

### Suma

```{r}
vector_numeros <- c(3, 5, 10)
2 + vector_numeros
```

- Podemos operar dos vectores de la misma longitud:

```{r}
vector1 <- c(1, 3, 5)
vector2 <- c(10, 20, 30)
vector1 + vector2 
```

- Podemos operar dos vectores de diferente longitud:

```{r}
vector3 <- c(10, 20)
vector4 <- c(10, 20, 30)
vector3 + vector4
```

### Potencia

```{r}
2^2
vector3^3
```

## Relacionales

### Igualdad

- En R la igualdad se denota con doble símbolo igual (==):

```{r}
50 == 40
```

- Podemos operar con la igualdad en vectores:

```{r}
vector_ejemplo <- c(40, 20, 2)
2 == vector_ejemplo
```

### Diferencia

- La diferencia en R se denota con un signo de exclamación de cierre y el igual (!=):

```{r}
3 != 3
```

- Tambien es posible operar de manera relacional con caracteres:

```{r}
texto <- c("A", "a", "b")
"a" != texto
```

### Pertece a o está dentro de

- En R podemos denotar el "pertene a" con %in%:

```{r}
numeros2 <- c(10, 20, 30, 40)
30 %in% numeros2
```

- Podemos operar sobre vectores de texto:

```{r}
texto2 <- c("a", "b", "c", "A")
"A" %in% texto2
```

### Mayor o igual

- En R el mayor o igual se denota con >=:

```{r}
ejemplo <- c(30, 40, 50, 10)
ejemplo >= 20
```

## Operadores lógicos

### Negación

```{r}
!2 == 2 # no es igual 2 a 2

!3 %in% c(2, 3, 5) # no pertenece el 3 al vector
```

### Disyunción (o)

```{r}
ejemplo2 <- c(1, 5, 10, 15, 20)
ejemplo2 > 10 | ejemplo2 < 5
```

### Conjunción (y)

```{r}
frutas1 <- c(80.5, 90, 50.4, 30)
frutas2 <- c(90, 30, 40, 20)

frutas1 > 80 | frutas2 > 80

frutas1 > 80 & frutas2 > 80
```

- ¿Qué pasa cuando utilizo el doble o y el doble y:

```{r}
frutas1 > 80 || frutas2 > 80

frutas1 > 80 && frutas2 > 80
```

# Funciones

```{r}
str(iris)
```

## Cargar biblioteca

```{r}
library(tidyverse)
```

# Vectores

- Almacenan información de un sólo tipo.
- Cuando mezclamos dos tipos de datos (numéricos y texto) R automáticamente coerciona al tipo de dato más sencillo.

## Coerción implícita

```{r}
mezcla1 <- c(1, 20, "a")
class(mezcla1)
mezcla1
```

## Coerción explícita

- Forzar la conversión de un tipo de dato a otro a través de las funciones con el prefijo **as.**

```{r}
mezcla2 <- as.numeric(mezcla1)
class(mezcla2)
mezcla2
```

- Otro ejemplo de coerción explícita:

```{r}
peso <- c("1.5", "2.50", "3.50", "4.5", "5")
class(peso)

peso2 <- as.numeric(peso)
peso2
class(peso2)

problema <- "2,2"
as.numeric(problema)
```

## Funciones sobre vectores

```{r}
vector1 <- c(1, 4, 20, 50, 20)
length(vector1)
vector1
```

- Los vectores tienen el atributo nombre:

```{r}
vector_nombrado <- c(elemento1 = 20, elemento2 = 50)
names(vector_nombrado)
vector_nombrado
```

- También puedo editar los nombres de un vector:

```{r}
names(vector_nombrado) <- c("variable1", "variable2")
vector_nombrado
```

## Secuencias de datos

- Los dos puntos (:) permiten crear secuencias de datos numéricos:

```{r}
secuencia1 <- c(1:10)
secuencia1
```

- Podemos crear secuencias numéricas controladas con la función seq():

```{r}
secuencia2 <- seq(from = 20, to = 50, by = 2)
secuencia2
```

- Podemos controlar la longitud de la secuencia:

```{r}
seq(from = 1, to = 20, length.out = 30)
```

## Repeticiones

```{r}
rep(1, 20)

rep("a", 5)

rep(c("A", "B", "C", "D"), 12)

rep(c("T1", "T2"), each = 12)
```

## Concatención (unión) de texto

```{r}
paste("T", 1, sep = "-")

mi_nombre <- "Edimer"

paste0("Hola ", mi_nombre, " ¿Cómo estás?")

paste0("T", 1:12)
```

## Subconjutos

- Podemos extraer subconjuntos a través de los corchetes [] haciendo uso del índice o nombre del elemento:

```{r}
vector5 <- c(10, 20, 30)
vector5[2]

vector5[c(1, 3)]

vector6 <- c(1, 10, 20, 30, 40, 50)
vector6[c(3:6)]

vector7 <- c(v1 = 3, v2 = 4, v3 = 5)
vector7["v2"]
```

- Podemos extraer subconjuntos con base en una condición:

```{r}
vector8 <- c(1, 2, 3, 6, 10, 50, 100, 500, 1000)

vector8[vector8 > 100]
```

- Ejemplo:

```{r}
vector9 <- c(10, 20)
vector9[c(FALSE, TRUE)]
```

- Vamos a extraer los valores que superan el promedio de vector8:

```{r}
promedio <- mean(vector8)
vector8[vector8 > promedio]
```

```{r}
vector8[!vector8 > promedio]
```

- Otro ejemplo:

```{r}
departamentos <- c("antioquia", "valle", "nariño", "caldas")
hurtos <- c(40, 50, 20, 15)

departamentos[hurtos > 20]
```

# Manejo de texto con stringr

## Concatenación de texto

```{r}
library(stringr)
mi_nombre <- "Edimer"
str_c("Hola ", mi_nombre, " ¿Cómo estás?")
```

- Concatenar dos vectores de texto:

```{r}
deptos <- c("Antioquia", "Santander")
mpios <- c("Medellín", "Bucaramanga")
str_c(deptos, "-", mpios)
```

## Conteo de caracteres

```{r}
texto4 <- "Este es mi texto"
str_length(texto4)

texto5 <- c("Mi texto", "antioquia", "cali")
str_length(texto5)
```

## Reemplazando texto

```{r}
numero_malo <- c("2,20")
as.numeric(numero_malo)

numero_bien <- str_replace_all(string = numero_malo,
                               pattern = ",",
                               replacement = ".")

as.numeric(numero_bien)
```

## Ordenar cadenas de texto

```{r}
texto6 <- c("1Felipe", "3Ramiro", "4Mauricio", "1Dora", "2Paola")
str_sort(texto6, decreasing = TRUE)

str_sort(texto6, decreasing = FALSE) # por defecto
```

## De mayúsculas a minúsculas

```{r}
texto7 <- c("EDIMER", "UNIVERSIDAD", "R")
str_to_lower(texto7)

texto8 <- c("Perro", "perro", "PERRO")
str_to_lower(texto8)
```

## De minúsculas a mayúsculas

```{r}
texto9 <- c("antioquia", "santander", "cundinamarca")
str_to_upper(texto9)
```

## Primera letra en mayúscula

```{r}
texto10 <- c("ANTIOQUIA", "valle del cauca")
str_to_sentence(texto10)
```

## Primeras letras en mayúscula

```{r}
str_to_title(texto10)
```

## Eliminar espacios en blanco

```{r}
texto12 <- c("Raza A", "Raza A      ", "     Raza A")
str_trim(texto12)
```

# Matrices

- Podemos crear una matriz con la función matrix():

```{r}
matriz1 <- matrix(c(2, 3, 4, 5))
matriz1
```

- Podemos controlar las dimensiones de la matriz:

```{r}
matriz2 <- matrix(c(10, 20, 30, 40), nrow = 2, ncol = 2)
matriz2
```

- Podemos controlar que sean generadas por filas:

```{r}
matriz3 <- matrix(c(100, 200, 300, 400), nrow = 2, ncol = 2,
                  byrow = TRUE)
matriz3
```

- Podemos consultar las dimensiones con la función dim():

```{r}
dim(matriz3)
dim(matriz1)
dim(iris) # sobre una base de datos
```

- Consultar sólo el número de filas con nrow():

```{r}
nrow(matriz1)
nrow(iris) # sobre una base de datos
```

- Consultar el número de columnas con ncol():

```{r}
ncol(matriz1)
ncol(iris) # sobre una base de datos
```

## Subconjuntos

- Ejercicio obtener primera columna:

```{r}
matriz4 <- matrix(c(1, 5, 50, 500, 200, 1000), ncol = 2, nrow = 3)
matriz4

matriz4[ , 1] # primera columna
matriz4[1, ] # primera fila

matriz4[c(1, 3), 2] # Queremos el 500 y el 1000
```

## Operaciones con matrices

```{r}
matriz5 <- matrix(c(10, 20, 30, 40), ncol = 2)
matriz5 + 1 # suma con escalar
matriz5 * 5 # multiplicación con escalar

matriz6 <- matrix(c(2, 4, 6, 8), ncol = 2)
matriz6

matriz5 %*% matriz6 # multiplicación de matrices

diag(matriz5) # diagonal
solve(matriz5) # inversa
```

# Listas

```{r}
lista1 <- list(1, "A", TRUE)
lista1
```

- Las listas tienen el atributo nombre:

```{r}
lista2 <- list(p1 = "Universidad", p2 = 1234, p3 = c(TRUE, FALSE))
lista2
```

- Consultando el elemento 3 de la lista2:

```{r}
lista2$p3
class(lista2)
class(lista2$p3)

lista2[3]
lista2[[3]]

lista2$p3[2]
```

# Agregando nuevos datos

## Vector

```{r}
vector8 <- c(1, 20, 50)
vector8[4] <- 100
vector8
```

## Matriz

```{r}
matriz7 <- matrix(c(10, 20, 30, 40), ncol = 2)
matriz7

# Agregando columnas
matriz8 <- cbind(matriz7, c(100, 200))
matriz8

# Agregando filas
rbind(matriz7, c(1000, 2000))

# Agregando filas de otra naturaleza
rbind(matriz7, c("A", "B"))
```

## Listas

```{r}
lista3 <- list(e1 = 1, e2 = "A", e3 = TRUE)
lista3

# Agregando nuevo dato
lista3$e4 <- NA
lista3

names(lista3)
length(lista3)
```

# Bases de datos

## Importación

### Formatos planos

- Separados por comas:

```{r}
library(readr)
datos_comas <- read_csv(file = "encuesta_comas.csv")
datos_comas
```

- Separados por punto y comas:

```{r}
datos_pcomas <- read_csv2(file = "encuesta_puntocomas.csv")
```

- Separados por tabulaciones (espacios en blanco):

```{r}
datos_tabulacion <- read_tsv(file = "encuesta_espacios.csv")
```

- Separados por otro delimitador (en este caso es /):

```{r}
datos_otro <- read_delim(file = "encuesta_delim.csv", delim = "/")
```

### Archivos de excel

```{r}
library(readxl)
datos_excel <- read_excel(path = "encuesta_excel.xlsx")
datos_excel
```

- Consultar los nombres de las hojas del archivo de excel:

```{r}
excel_sheets(path = "encuesta_excel.xlsx")
```

## Funciones básicas

```{r}
names(datos_excel) # nombres de columnas

nrow(datos_excel) # número de filas

ncol(datos_excel) # número de columnas

dim(datos_excel) # Dimensiones (filas y columnas)

class(datos_excel$promedio_academico) # Clase de una variable

library(dplyr)
glimpse(datos_excel)
```

## Subconjuntos

- Filtrar sólo las primeras 20 filas de la base de datos:

```{r}
datos_excel[1:20, ]
```

- Filtrar sólo las filas 1, 10, 20 y 30

```{r}
datos_excel[c(1, 10, 20, 30), ]
```

- Filtrar sólo las filas 1 y 10 con las variables color_favorito y horas_dormir.

```{r}
datos_excel[c(1, 10), c("color_favorito", "horas_dormir")]
```

- Filtrar datos de estudiantes que superan la mediana del promedio académico:

```{r}
mediana <- median(datos_excel$promedio_academico)

datos_excel[datos_excel$promedio_academico > mediana, ]
```

- Filtrar datos de estudiantes que superan la mediana del promedio acádemico y cuyo color favorito es el Negro

```{r}
datos_excel[datos_excel$promedio_academico > mediana
            & datos_excel$color_favorito == "Negro", ]
```

## Data frame manual

```{r}
departamento <- c("Antioquia", "Santander", "Cundinamarca")
casos_covid <- c(1234, 2340, 4580)

mis_datos <- data.frame(depto = departamento, covid = casos_covid)
mis_datos
```
