==> temas/Guides/examples_rproject/03-Manejo-datos-dplyr.Rmd <==
---
title: "Manejo de datos con R (dplyr)"
author: "Edimer David Jaramillo"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: yes
    toc_float: yes
    number_sections: yes
    theme: cerulean
    highlight: pygments
    df_print: paged
    code_folding: show
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```

# Lectura de datos

```{r}
library(tidyverse)
datos <- read_csv("ejemplo.csv")
datos
```

# Operaciones por filas

## filter()

- Ejemplo: filtrando sólo el año (year) 2000.

```{r}
# Con corchetes: datos[datos$year == 2000, ]
datos %>% 
  filter(year == 2000)
```

- Ejemplo: filtrando sólo el año 2000 y homicidios mayores a 70.

```{r}
datos %>% 
  filter(year == 2000 & homicidio > 70)  
```

## slice()

- La diferencia de slice() con filter() es que la primera filtra filas en  función de la posición. La segunda filtra filas en función de una o más condiciones.
- Ejemplo: filtramos sólo las filas 1, 10 y 20.

```{r}
datos %>% 
  slice(1, 10, 20)
```

## arrange()

- La función arrange() permite ordenar las filas de manera ascendente (por defecto) o descendente.
- Ejemeplo: ordenando las filas por la variable "homicidio".

```{r}
datos %>% 
  arrange(homicidio)
```

- Ejemplo: ordenando las filas por la variable "homicidio" de forma descendente. Se utiliza la función "desc()".

```{r}
datos %>% 
  arrange(desc(homicidio))
```

## distinct()

- Permite filtrar filas únicas. Por defecto sólo selecciona la columna implicada en el filtro de filas únicas o distintas.
- Ejemplo: filtrando municipios diferentes o únicos

```{r}
datos %>% 
  distinct(Municipio)
```

- Podemos mantener las demás columnas a través del argumento .keep_all = TRUE.

```{r}
datos %>% 
  distinct(Municipio, .keep_all = TRUE)
```

# Operaciones por columnas

## rename()

- Permite renombrar una o más variables.
- Ejemplo: cambiar el nombre del municipio a minúscula.

```{r}
# Ejemplo con la función names()
prueba <- datos
names(prueba) <- c("municipio", "year", "homicidio")
prueba
```

- Ahora utilizamos rename(): al lado derecho del igual está el nombre original y al lado izquierdo el nuevo nombre.

```{r}
datos %>% 
  rename(municipio = Municipio)
```

## set_names()

- Permite cambiar el nombre a todas las columnas a través de un vector de caracteres. Similar a lo que hace names(), sin embargo, set_names() es compatible con el operador de tubería.

```{r}
datos %>% 
  set_names(c("var1", "var2", "var3"))
```

## select()

- Permite seleccionar columnas a través del nombre o la posición.

```{r}
datos %>% 
  select(year, homicidio)
```

- Este código arroja el mismo resultado anterior:

```{r}
datos %>% 
  select(-Municipio)
```

- Seleccionar sólo las variables numéricas:

```{r}
datos %>% 
  select(is.numeric)
```

- Seleccionar las variables tipo caracter y además el homicidio:

```{r}
datos %>% 
  select(is.character, homicidio)
```

- También se puede usar el número de la columna:

```{r}
datos %>% 
  select(1, 3)
```

# mutate()

- mutate() permite editar columnas o crear nuevas variables.
- Ejemplo: crear una nueva columna de nombre "pais" con la etiqueta "Colombia"

```{r}
datos %>% 
  mutate(pais = "Colombia")
```

- Se aplica coerción sobre la variable "Municipio", de character a factor:

```{r}
datos %>% 
  mutate(Municipio = as.factor(Municipio))
```

- Se resta el número 10 a la columna homicidio:

```{r}
datos %>% 
  mutate(homicidio2 = homicidio - 10)
```

- Puedo aplicar más de una edición o incluso crear y editar en la misma sentencia:

```{r}
datos %>% 
  mutate(Municipio = as.factor(Municipio),
         homicidio2 = homicidio - 10)
```

# Resumen de datos

## count()

- Esta función permite contar observaciones (filas) en función de una o más variables

```{r}
datos %>% 
  count(year)
```

- Conteo de registros (filas) por municipio y año

```{r}
datos %>% 
  count(Municipio, year, name = "total")
```

## group_by() + summarise()

- Ejemplo: obtener resumen numérico por año que conste de promedio, desviación estándar, mínimo y máximo para la tasa de homicidio.

```{r}
datos %>% 
  group_by(year) %>% 
  summarise(
    promedio_homicidio = mean(homicidio),
    desv_homicidio = sd(homicidio),
    minimo_homicidio = min(homicidio),
    maximo_homicidio = max(homicidio)
  ) 
```

## ungroup()

- Después de utilizar group_by() y summarise() se recomienda usar la función ungroup().

```{r}
df_agrupada <- datos %>% 
  group_by(year)

# Consultando la clase de df_agrupada
class(df_agrupada)

# Tabla desagrupada
df_desagrupada <- df_agrupada %>% 
  ungroup()

# Consultando la clase de df_desagrupada
class(df_desagrupada)
```

- Se recomienda utilizar ungroup() después del group_by() + summarise():

```{r}
datos %>% 
  group_by(Municipio) %>% 
  summarise(promedio = mean(homicidio)) %>% 
  ungroup()
```

## group_by() + mutate()

```{r}
datos %>% 
  group_by(Municipio) %>% 
  mutate(promedio = mean(homicidio)) %>% 
  ungroup()
```
